updated models, views, subscriptions
Daniel and I have spent the day getting the form to subscribe and save the data lots have been done so take time to view the changes, TODO still is the validation of the information being passed in and to update current data form. add some more helper text. small issue on the decorator as accessing wrong model this will be updated tomorrow

Co-Authored-By: minlopalis <52741304+minlopalis@users.noreply.github.com>